css/css-masking/clip-path/clip-path-path-with-zoom-hittest.html

<!DOCTYPE html>
<html>
<head>
  <title>CSS Masking: Test clip-path property when the page is zoomed</title>
  <link rel="author" title="Noam Rosenthal" href="mailto:noam@webkit.org">
  <link rel="help" href="https://drafts.csswg.org/css-shapes-2/#funcdef-path">
  <link rel="match" href="reference/clip-path-path-with-zoom-ref.html">
  <meta name="assert" content="The path gets zoomed together with the content">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<style>
  #triangle {
    width: 100px;
    height: 100px;
    background-color: green;
    clip-path: path(nonzero, 'M0 0, L100 0, L0 100, L 0 0');
    zoom: 2;
  }
</style>
<body>
  <div id="triangle"></div>
  <output id="log"></output>
  <script>
    const triangle = document.getElementById('triangle')
    test(() => {
      assert_equals(document.elementFromPoint(20, 20).id, 'triangle')
      assert_equals(document.elementFromPoint(150, 20).id, 'triangle')
      assert_equals(document.elementFromPoint(180, 180).tagName, 'BODY')
    }, 'clip-path: path() hit-test takes zoom into account');

  </script>
</body>
</html>
